+++ routers/index.js
@@ -1,15 +1,10 @@
 const express = require("express");
+const router = express.Router();
 
-//const mainRouter = require("./main.js");
-//const userRouter = require("./user.js");
 const taskRouter = require("./task.js");
 const authRouter = require("./auth.js");
 
-const router = express.Router();
-
 router.use("/", taskRouter);
-//router.use("/user", userRouter);
 router.use("/auth", authRouter);
-//router.use(mainRouter);
 
 module.exports = router;
